// packages/client/src/store/home/actionTypes.ts
export enum HomeActionType {
  GetGames = 'HOME.GET_GAMES',
  SetGames = 'HOME.SET_GAMES',
  setGamesForSearch = 'GAMES.SET_SEARCH_DATA',
  getGamesForSearch = 'GAMES.GET_SEARCH_DATA',
  FetchBanners = 'HOME.FETCH_BANNERS',
  BannersFetched = 'HOME.BANNERS_FETCHED',
  SetProviders = 'HOME.SET_GAME_PROVIDERS',
  GetProviders = 'HOME.GET_GAME_PROVIDERS',
  SetRecommendationsGames = 'HOME.SET_RECOMMENDATIONS_GAMES',
  GetRecommendationsGames = 'HOME.GET_RECOMMENDATIONS_GAMES',
  SetAssets = 'HOME.SET_ASSETS',
  GetAssets = 'HOME.GET_ASSETS',
}